// src/types/Character.type.ts
//attributes
export interface Character {
    name: string;
    height: string;
    gender: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    homeworld: string;
    starships: string[];
    films: string[];
    species: string[];
    vehicles: string[];
    planets: string[];
    url: string;
}